dev: Convert core.git into a package
_more: The module started to grow big, makes it harder to maintain.
