+++ test/unit/core/git/test_ignore.py
@@ -11,7 +11,7 @@ FAKE_GIT_IGNORE = '\n'.join([
 ])
 
 
-@patch('peltak.core.git.open', mock_open(read_data='\n'.join([
+@patch('peltak.core.git.util.open', mock_open(read_data='\n'.join([
     'pattern1',
     'pattern2',
     'pattern3',
@@ -26,7 +26,7 @@ def test_returns_all_patterns(app_conf):
     util.cached_result.clear(git.ignore)
 
 
-@patch('peltak.core.git.open', mock_open(read_data='\n'.join([
+@patch('peltak.core.git.util.open', mock_open(read_data='\n'.join([
     'pattern1',
     ' pattern2   \t',
     '\tpattern3',
@@ -41,7 +41,7 @@ def test_strips_whitespace(app_conf):
     util.cached_result.clear(git.ignore)
 
 
-@patch('peltak.core.git.open', mock_open(read_data='\n'.join([
+@patch('peltak.core.git.util.open', mock_open(read_data='\n'.join([
     'pattern1',
     ''
     ' pattern2   \t',
@@ -59,7 +59,7 @@ def test_skips_empty_lines(app_conf):
     util.cached_result.clear(git.ignore)
 
 
-@patch('peltak.core.git.open', mock_open(read_data=b'\n'.join([
+@patch('peltak.core.git.util.open', mock_open(read_data=b'\n'.join([
     b'pattern1',
     b''
     b' pattern2   \t',
